PrevailingTrend/第一层模块/公司名字列表/requirements_new.txt
# 数据库连接
pymysql>=1.0.0
sqlalchemy>=1.4.0

# 数据处理
pandas>=1.3.0
numpy>=1.21.0

# HTTP请求
requests>=2.25.0

# Web框架
flask>=2.0.0
flask-cors>=3.0.10

# 配置管理
python-dotenv>=0.19.0

# 测试框架
pytest>=6.0.0
pytest-cov>=2.12.0

# 代码质量
black>=21.0.0
flake8>=3.9.0

# 日志管理
loguru>=0.5.0

# 数据验证
pydantic>=1.8.0

# 异步支持
aiohttp>=3.8.0
asyncio>=3.4.3

# 万得Wind接口 (可选)
# WindPy>=0.8.0

# 免费数据源
tushare>=1.2.0
akshare>=1.0.0

# 日期时间处理
python-dateutil>=2.8.0

# 正则表达式
regex>=2021.0.0

# 缓存支持
redis>=4.0.0

# 监控和指标
prometheus-client>=0.12.0 